z3str3: refactoring, move legacy model construction code into theory_str_mc
